Fix observation type

+++ sacking/policy.py
@@ -89,7 +89,7 @@ class GaussianPolicy(nn.Module):
             -> np.ndarray:
         """Choose action from policy for one observation."""
         with torch.no_grad():
-            pt_obs = torch.from_numpy(observation).unsqueeze(0)
+            pt_obs = torch.from_numpy(observation).unsqueeze(0).float()
             if mode == 'sample':
                 pt_action, _ = self.sample_action(pt_obs)
             elif mode == 'best':
